refactor: split executions FS chart and switch
