+++ ui/src/components/dashboard/components/charts/executions/Bar.vue
@@ -30,10 +30,11 @@
             </div>
         </div>
 
-        <Bar
+        <BarChart
             v-if="total > 0"
-            :data="parsedData"
-            :options="options"
+            :data="data"
+            :total="total"
+            :duration="duration"
             :plugins="[barLegend]"
             class="tall"
         />
@@ -43,26 +44,19 @@
 </template>
 
 <script setup>
-    import {computed, ref, onMounted, onUnmounted} from "vue";
+    import {ref} from "vue";
     import {useI18n} from "vue-i18n";
-
-    import moment from "moment";
-    import {Bar} from "vue-chartjs";
 
     import {barLegend} from "../legend.js";
 
-    import Utils from "../../../../../utils/utils.js";
-    import {defaultConfig, getFormat} from "../../../../../utils/charts.js";
-    import {getScheme} from "../../../../../utils/scheme.js";
-
     import NoData from "../../../../layout/NoData.vue";
 
-    import Check from "vue-material-design-icons/Check.vue";
+    import BarChart from "./BarChart.vue";
 
     const {t} = useI18n({useScope: "global"});
-    const isSmallScreen = ref(window.innerWidth < 610);
+    const duration = ref(true);
 
-    const props = defineProps({
+    defineProps({
         data: {
             type: Object,
             required: true,
@@ -72,160 +66,6 @@
             required: true,
         },
     });
-
-    const parsedData = computed(() => {
-        let datasets = props.data.reduce(function (accumulator, value) {
-            Object.keys(value.executionCounts).forEach(function (state) {
-                if (accumulator[state] === undefined) {
-                    accumulator[state] = {
-                        label: state,
-                        backgroundColor: getScheme(state),
-                        yAxisID: "y",
-                        data: [],
-                    };
-                }
-
-                accumulator[state].data.push(value.executionCounts[state]);
-            });
-
-            return accumulator;
-        }, Object.create(null));
-
-        return {
-            labels: props.data.map((r) =>
-                moment(r.startDate).format(getFormat(r.groupBy)),
-            ),
-            datasets: duration.value
-                ? [
-                    {
-                        type: "line",
-                        label: t("duration"),
-                        fill: false,
-                        pointRadius: 0,
-                        borderWidth: 0.75,
-                        borderColor: "#A2CDFF",
-                        yAxisID: "yB",
-                        data: props.data.map((value) => {
-                            return value.duration.avg === 0
-                                ? 0
-                                : Utils.duration(value.duration.avg);
-                        }),
-                    },
-                    ...Object.values(datasets),
-                ]
-                : Object.values(datasets),
-        };
-    });
-
-    onMounted(() => {
-        const handleResize = () => {
-            isSmallScreen.value = window.innerWidth < 610;
-        };
-        window.addEventListener("resize", handleResize);
-
-        onUnmounted(() => {
-            window.removeEventListener("resize", handleResize);
-        });
-    });
-
-    const options = computed(() =>
-        defaultConfig({
-            barThickness: isSmallScreen.value ? 8 : 12,
-            skipNull: true,
-            borderSkipped: false,
-            borderColor: "transparent",
-            borderWidth: 2,
-            plugins: {
-                barLegend: {
-                    containerID: "executions",
-                },
-                tooltip: {
-                    enabled: true,
-                    filter: (value) => value.raw,
-                    callbacks: {
-                        label: (value) => {
-                            const {label, yAxisID} = value.dataset;
-                            return `${label.toLowerCase().capitalize()}: ${value.raw}${yAxisID === "yB" ? "s" : ""}`;
-                        },
-                    },
-                },
-            },
-            scales: {
-                x: {
-                    title: {
-                        display: true,
-                        text: t("date"),
-                    },
-                    grid: {
-                        display: false,
-                    },
-                    position: "bottom",
-                    display: true,
-                    stacked: true,
-                    ticks: {
-                        maxTicksLimit: isSmallScreen.value ? 5 : 8,
-                        callback: function (value) {
-                            const label = this.getLabelForValue(value);
-
-                            if (
-                                moment(label, ["h:mm A", "HH:mm"], true).isValid()
-                            ) {
-                                // Handle time strings like "1:15 PM" or "13:15"
-                                return moment(label, ["h:mm A", "HH:mm"]).format(
-                                    "h:mm A",
-                                );
-                            } else if (moment(new Date(label)).isValid()) {
-                                // Handle date strings
-                                const date = moment(new Date(label));
-                                const isCurrentYear =
-                                    date.year() === moment().year();
-                                return date.format(
-                                    isCurrentYear ? "MM/DD" : "MM/DD/YY",
-                                );
-                            }
-
-                            // Return the label as-is if it's neither a valid date nor time
-                            return label;
-                        },
-                    },
-                },
-                y: {
-                    title: {
-                        display: !isSmallScreen.value,
-                        text: t("executions"),
-                    },
-                    grid: {
-                        display: false,
-                    },
-                    display: true,
-                    position: "left",
-                    stacked: true,
-                    ticks: {
-                        maxTicksLimit: isSmallScreen.value ? 5 : 8,
-                    },
-                },
-                yB: {
-                    title: {
-                        display: duration.value && !isSmallScreen.value,
-                        text: t("duration"),
-                    },
-                    grid: {
-                        display: false,
-                    },
-                    display: duration.value,
-                    position: "right",
-                    ticks: {
-                        maxTicksLimit: isSmallScreen.value ? 5 : 8,
-                        callback: function (value) {
-                            return `${this.getLabelForValue(value)}s`;
-                        },
-                    },
-                },
-            },
-        }),
-    );
-
-    const duration = ref(true);
 </script>
 
 <style lang="scss" scoped>
